feat(segmentation): Enhance labelmap functionality with new updates
- Added a button to the toolbar for adding labelmap representations to the viewport.
- Refactored segmentation data modification logic to utilize new helper functions for volume and stack updates.
- Removed outdated timeout logic for adding labelmap representations.
- Introduced new files for handling volume and stack labelmap updates, improving code organization and maintainability.
