+++ packages/tools/src/eventListeners/segmentation/labelmap/onLabelmapSegmentationDataModified.ts
@@ -1,16 +1,15 @@
 import {
-  cache,
-  utilities as csUtils,
   VolumeViewport,
   getEnabledElementByViewportId,
   StackViewport,
 } from '@cornerstonejs/core';
 
-import * as SegmentationState from '../../../stateManagement/segmentation/segmentationState';
 import type { SegmentationDataModifiedEventType } from '../../../types/EventTypes';
-import type { LabelmapSegmentationDataVolume } from '../../../types/LabelmapTypes';
 import { SegmentationRepresentations } from '../../../enums';
-import { getLabelmapActorEntries } from '../../../stateManagement/segmentation/helpers/getSegmentationActor';
+import { performVolumeLabelmapUpdate } from './performVolumeLabelmapUpdate';
+import { performStackLabelmapUpdate } from './performStackLabelmapUpdate';
+import { getSegmentation } from '../../../stateManagement/segmentation/getSegmentation';
+import { getViewportIdsWithSegmentation } from '../../../stateManagement/segmentation/getViewportIdsWithSegmentation';
 
 /** A callback function that is called when the segmentation data is modified which
  *  often is as a result of tool interactions e.g., scissors, eraser, etc.
@@ -20,11 +19,9 @@ const onLabelmapSegmentationDataModified = function (
 ): void {
   const { segmentationId, modifiedSlicesToUse } = evt.detail;
 
-  const { representationData } =
-    SegmentationState.getSegmentation(segmentationId);
+  const { representationData } = getSegmentation(segmentationId);
 
-  const viewportIds =
-    SegmentationState.getViewportIdsWithSegmentation(segmentationId);
+  const viewportIds = getViewportIdsWithSegmentation(segmentationId);
 
   const hasVolumeViewport = viewportIds.some((viewportId) => {
     const { viewport } = getEnabledElementByViewportId(viewportId);
@@ -61,110 +58,6 @@ const onLabelmapSegmentationDataModified = function (
       });
     }
   });
-
-  // Todo: i don't think we need this anymore
-  // if (
-  //   'stack' in labelmapRepresentationData &&
-  //   'volumeId' in labelmapRepresentationData
-  // ) {
-  //   // we need to take away the modifiedSlicesToUse from the stack
-  //   // and update the volume for all the slices
-  //   modifiedSlices = [];
-  // }
 };
-
-function performVolumeLabelmapUpdate({
-  modifiedSlicesToUse,
-  representationData,
-  type,
-}) {
-  const segmentationVolume = cache.getVolume(
-    (representationData[type] as LabelmapSegmentationDataVolume).volumeId
-  );
-
-  if (!segmentationVolume) {
-    console.warn('segmentation not found in cache');
-    return;
-  }
-
-  const { imageData, vtkOpenGLTexture } = segmentationVolume;
-
-  // Update the texture for the volume in the GPU
-  let slicesToUpdate;
-  if (modifiedSlicesToUse?.length > 0) {
-    slicesToUpdate = modifiedSlicesToUse;
-  } else {
-    const numSlices = imageData.getDimensions()[2];
-    slicesToUpdate = [...Array(numSlices).keys()];
-  }
-
-  slicesToUpdate.forEach((i) => {
-    vtkOpenGLTexture.setUpdatedFrame(i);
-  });
-
-  // Trigger modified on the imageData to update the image
-  // this is the start of the rendering pipeline for updating the texture
-  // to the gpu
-  imageData.modified();
-}
-
-function performStackLabelmapUpdate({ viewportIds, segmentationId }) {
-  viewportIds.forEach((viewportId) => {
-    let representations = SegmentationState.getSegmentationRepresentations(
-      viewportId,
-      { segmentationId }
-    );
-
-    representations = representations.filter(
-      (representation) =>
-        representation.type === SegmentationRepresentations.Labelmap
-    );
-
-    representations.forEach((representation) => {
-      if (representation.segmentationId !== segmentationId) {
-        return;
-      }
-
-      const enabledElement = getEnabledElementByViewportId(viewportId);
-
-      if (!enabledElement) {
-        return;
-      }
-
-      const { viewport } = enabledElement;
-
-      if (viewport instanceof VolumeViewport) {
-        return;
-      }
-
-      const actorEntries = getLabelmapActorEntries(viewportId, segmentationId);
-
-      if (!actorEntries?.length) {
-        return;
-      }
-
-      actorEntries.forEach((actorEntry, i) => {
-        const segImageData = actorEntry.actor.getMapper().getInputData();
-
-        const currentSegmentationImageIds =
-          SegmentationState.getCurrentLabelmapImageIdsForViewport(
-            viewportId,
-            segmentationId
-          );
-
-        const segmentationImage = cache.getImage(
-          currentSegmentationImageIds[i]
-        );
-        segImageData.modified();
-
-        // update the cache with the new image data
-        csUtils.updateVTKImageDataWithCornerstoneImage(
-          segImageData,
-          segmentationImage
-        );
-      });
-    });
-  });
-}
 
 export default onLabelmapSegmentationDataModified;
